// src/types/nutrition.ts
export type ServingInfo = {
  amount: number
  unit: string
  gramWeight: number
}
// ...
export type USDAFoodPortion = {
  id: number
  amount: number
  gramWeight: number
  modifier?: string
  portionDescription?: string
  measureUnit?: {
    id: number
    name: string
    abbreviation: string
  }
}